float percents

+++ src/main/java/ru/evotor/framework/inventory/product/category/entertainment/StrongAlcohol.kt
@@ -29,7 +29,7 @@ data class StrongAlcohol internal constructor(
         override val vatRate: VatRate,
         override val quantity: Quantity,
         override val tareVolume: AmountOfLiters,
-        override val alcoholPercentage: BigDecimal?,
+        override val alcoholPercentage: Float?,
         override val description: String?
 ) : Product(), AlcoholProduct, ExcisableProduct {
 
@@ -49,7 +49,7 @@ data class StrongAlcohol internal constructor(
         val vatRate = addFieldFilter<VatRate>(InventoryContract.Product.VAT_RATE)
         val quantity = addInnerFilterBuilder(Quantity.Filter<Query, SortOrder, StrongAlcohol>())
         val tareVolume = addFieldFilter<AmountOfLiters>(InventoryContract.Product.TARE_VOLUME)
-        val alcoholPercentage = addFieldFilter<BigDecimal?>(InventoryContract.Product.ALCOHOL_PERCENTAGE)
+        val alcoholPercentage = addFieldFilter<Float?>(InventoryContract.Product.ALCOHOL_PERCENTAGE)
         val description = addFieldFilter<String?>(InventoryContract.Product.DESCRIPTION)
 
         class SortOrder : FilterBuilder.SortOrder<SortOrder>() {
